% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{datafolder_update}
\alias{datafolder_update}
\title{Save data folder content list to sync to git}
\usage{
datafolder_update(docs_folder = "docs", data_folder = "data")
}
\arguments{
\item{docs_folder}{subfolder where the data_folder_content.csv file is located. Currently cannot be the working directory root.}

\item{data_folder}{data folder path}
}
\value{
Nothing, writes docs/data_folder_content.csv directly. This csv has two columns, \code{filepath}, and \code{md5}
}
\description{
Creates the docs/data_folder_content.csv file which can be synced to git, and used by \code{\link{datafolder_check}} to make sure the
local data folder is up-to-date with the rest of the code
}
\details{
The md5 checksum is created using \code{openssl::md5(file(<filename>))}
}
\seealso{
\code{\link{datafolder_check}}
}
